+++ src/www/app/pages/reference/employeelist.php
@@ -24,6 +24,7 @@ class EmployeeList extends \App\Pages\Base
 
     public function __construct() {
         parent::__construct();
+        if(false ==\App\ACL::checkShowRef('EmployeeList'))return;       
 
         $this->add(new Panel('employeetable'))->setVisible(true);
         $this->employeetable->add(new DataView('employeelist', new EDS('\App\Entity\Employee'), $this, 'employeelistOnRow'))->Reload();
@@ -48,6 +49,8 @@ class EmployeeList extends \App\Pages\Base
     }
 
     public function deleteOnClick($sender) {
+        if(false ==\App\ACL::checkEditRef('EmployeeList'))return;       
+        
         Employee::delete($sender->owner->getDataItem()->employee_id);
         $this->employeetable->employeelist->Reload();
     }
@@ -73,7 +76,8 @@ class EmployeeList extends \App\Pages\Base
     }
 
     public function saveOnClick($sender) {
-
+       if(false ==\App\ACL::checkEditRef('EmployeeList'))return;       
+ 
         $login = trim($this->employeedetail->editlogin->getText());
         if (strlen($login) > 0) {
             if ($login == "admin") {
